+++ source/Engine/Core/Window.h
@@ -86,8 +86,8 @@ public:
    // Implement Bounded
    uint32_t GetX() const override { return 0; }
    uint32_t GetY() const override { return 0; }
-   uint32_t GetWidth() const override { return mOptions.width; }
-   uint32_t GetHeight() const override { return mOptions.height; }
+   uint32_t GetWidth() const override { return (uint32_t)mOptions.width; }
+   uint32_t GetHeight() const override { return (uint32_t)mOptions.height; }
 
 public:
    // Typical GLFW methods
